requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cssutils==1.0.2
diff-match-patch==20181111
Django==3.0.6
django-mathfilters==1.0.0
lxml==4.5.1
Pillow==7.1.2
pytz==2020.1
quill-delta==1.0.1
sqlparse==0.3.1
wincertstore==0.2
